Change int to uint

+++ MachineEmulator/Program.cs
@@ -8,7 +8,7 @@ var interruptDevice = new HardwareInterruptDevice();
 var ram = new RAM();
 for (var i = 0; i < machineCode.Count; i++)
 {
-    ram.SetDWord(0x408 + i * 4, machineCode[i]);
+    ram.SetDWord((ulong)(0x408 + i * 4), machineCode[i]);
 }
 
 var processor = new Processor(ram, interruptDevice);
